+++ .github/workflows/commitlint.yml
@@ -21,7 +21,7 @@ jobs:
           fetch-depth: 0
 
       - name: Setup Node and pnpm
-        uses: silverhand-io/actions-node-pnpm-run-steps@v1.2.3
+        uses: silverhand-io/actions-node-pnpm-run-steps@v2
 
       - name: Commitlint
         run: npx commitlint --from HEAD~${{ github.event.pull_request.commits }} --to HEAD
