+++ .npmrc
@@ -2,3 +2,5 @@
 public-hoist-pattern[]=@parcel/*
 public-hoist-pattern[]=postcss
 public-hoist-pattern[]=*eslint*
+public-hoist-pattern[]=process
+public-hoist-pattern[]=buffer
