+++ packages/browser-sample/package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0-beta.15",
   "license": "MIT",
   "private": true,
+  "source": "public/index.html",
   "scripts": {
     "precommit": "lint-staged",
     "start": "parcel public/index.html -p 3000",
@@ -18,13 +19,11 @@
     "@parcel/transformer-sass": "^2.8.3",
     "@silverhand/eslint-config": "^1.0.0",
     "@silverhand/eslint-config-react": "^1.0.0",
-    "buffer": "^6.0.3",
     "eslint": "^8.23.0",
     "lint-staged": "^13.0.0",
     "parcel": "^2.8.3",
     "postcss": "^8.4.6",
     "prettier": "^2.7.1",
-    "process": "^0.11.10",
     "stylelint": "^14.11.0",
     "typescript": "4.7.4"
   },
