+++ packages/js/src/core/sign-in.test.ts
@@ -1,5 +1,4 @@
-import { Prompt } from '../consts';
-import { UserScope } from '../utils/index.js';
+import { Prompt, UserScope } from '../consts';
 import { generateSignInUri } from './sign-in';
 
 const authorizationEndpoint = 'https://logto.dev/oidc/sign-in';
