+++ packages/react-sample/package.json
@@ -27,12 +27,14 @@
     "@types/react": "^17.0.39",
     "@types/react-dom": "^17.0.11",
     "@types/react-router-dom": "^5.3.3",
+    "buffer": "^6.0.3",
     "eslint": "^8.23.0",
     "lint-staged": "^13.0.0",
     "parcel": "^2.8.3",
     "postcss": "^8.4.6",
     "postcss-modules": "^5.0.0",
     "prettier": "^2.7.1",
+    "process": "^0.11.10",
     "stylelint": "^14.11.0",
     "typescript": "4.7.4"
   },
